add recyclerview to animal list page

+++ app/src/main/java/org/wit/cowcalendar/models/EventStore.kt
@@ -5,6 +5,6 @@ interface EventStore {
   fun create(event: EventModel)
   fun update(event: EventModel)
   fun delete(event: EventModel)
-  fun findById(id:Long) :EventModel?
+  fun findById(id:Int) :EventModel?
   fun clear ()
 }
